feat: Client exports dehydrate
