+++ packages/cozy-client/src/index.js
@@ -10,3 +10,4 @@ export { default as compose } from 'lodash/flow'
 export { QueryDefinition, MutationTypes, getDoctypeFromOperation } from './dsl'
 export { queryConnect, withClient } from './hoc'
 export { Association, HasManyAssociation } from './associations'
+export { dehydrate } from './helpers'
